Refactor MOTEUR layout
