+++ MOTEUR/compta/dashboard_widget.py
