+++ MOTEUR/scraping/scraping_widget.py
@@ -20,9 +20,9 @@ from PySide6.QtWidgets import (
     QComboBox,
 )
 
-from MOTEUR.scraping.image_scraper import download_images
-from MOTEUR.scraping.constants import IMAGES_DEFAULT_SELECTOR
-from MOTEUR.scraping.profile_manager import ProfileManager
+from .image_scraper import download_images
+from .constants import IMAGES_DEFAULT_SELECTOR
+from .profile_manager import ProfileManager
 
 
 class LogHandler(logging.Handler, QObject):
