+++ main.py
@@ -21,11 +21,11 @@ except ModuleNotFoundError:
     print("Install dependencies with pip install -r requirements.txt")
     sys.exit(1)
 
-from MOTEUR.scraping_widget import ScrapingImagesWidget
-from MOTEUR.achat_widget import AchatWidget
-from MOTEUR.vente_widget import VenteWidget
-from MOTEUR.profile_widget import ProfileWidget
-from MOTEUR.dashboard_widget import DashboardWidget
+from MOTEUR.scraping.scraping_widget import ScrapingImagesWidget
+from MOTEUR.compta.achat_widget import AchatWidget
+from MOTEUR.compta.vente_widget import VenteWidget
+from MOTEUR.scraping.profile_widget import ProfileWidget
+from MOTEUR.compta.dashboard_widget import DashboardWidget
 import subprocess
 
 BASE_DIR = Path(__file__).resolve().parent
